move error-boundary component to a separate file
