+++ src/components/error-boundary/error-boundary.jsx
@@ -1,0 +1,35 @@
+// libraries
+import React from "react";
+
+
+
+export default class ErrorBoundary extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { 
+      hasError: false 
+    };
+  };
+
+  static getDerivedStateFromError(error) {
+    return {hasError: true};
+  };
+
+  componentDidCatch(error, info) {
+    console.log("error:", error, info);
+  };
+
+  render() {
+    if (this.state.hasError) {
+      return (
+        <section>
+          <h1>Что-то пошло не так!</h1>
+          <p>
+            В приложении произошла ошибка. Пожалуйста, перезагрузите страницу.
+          </p>
+        </section>
+      );
+    }
+    return this.props.children;
+  };
+};
