+++ src/index.js
@@ -4,42 +4,10 @@ import ReactDOM from "react-dom";
 
 // components
 import App from "./components/app/app";
+import ErrorBoundary from "./components/error-boundary/error-boundary";
 
 // styles
 import "./index.css";
-
-
-
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { 
-      hasError: false 
-    };
-  };
-
-  static getDerivedStateFromError(error) {
-    return {hasError: true};
-  };
-
-  componentDidCatch(error, info) {
-    console.log("error:", error, info);
-  };
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <section>
-          <h1>Что-то пошло не так :(</h1>
-          <p>
-            В приложении произошла ошибка. Пожалуйста, перезагрузите страницу.
-          </p>
-        </section>
-      );
-    }
-    return this.props.children;
-  };
-};
